+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   - Fix formatting of titles where remix is delimited by a dash or an em-dash.
   - For some releases, handle track names where artist and title may be provided in
     reverse order.
+  - Address an issue where track names containing **` - `** _inside parentheses_ get
+    wrongly split into artist and title.
 
 ### Updated